+++ src/core/index.js
@@ -1,0 +1,5 @@
+const DBUtil = require('./utils/dbutil');
+
+module.exports = {
+    DBUtil
+}
