--- utils/index.js
@@ -1,5 +1,0 @@
-const DBUtil = require('./dbutil');
-
-module.exports = {
-    DBUtil
-}
